+++ oauth2/jwt.go
@@ -1,4 +1,4 @@
-package jwt
+package oauth2
 
 import (
 	"context"
@@ -6,11 +6,10 @@ import (
 	"time"
 
 	gojwt "github.com/dgrijalva/jwt-go"
-	"github.com/influxdata/chronograf"
 )
 
 // Test if JWT implements Authenticator
-var _ chronograf.Authenticator = &JWT{}
+var _ Authenticator = &JWT{}
 
 // JWT represents a javascript web token that can be validated or marshaled into string.
 type JWT struct {
@@ -45,7 +44,7 @@ func (c *Claims) Valid() error {
 }
 
 // Authenticate checks if the jwtToken is signed correctly and validates with Claims.
-func (j *JWT) Authenticate(ctx context.Context, jwtToken string) (chronograf.Principal, error) {
+func (j *JWT) Authenticate(ctx context.Context, jwtToken string) (Principal, error) {
 	gojwt.TimeFunc = j.Now
 
 	// Check for expected signing method.
@@ -62,27 +61,31 @@ func (j *JWT) Authenticate(ctx context.Context, jwtToken string) (chronograf.Pri
 	// 4. Check if subject is not empty
 	token, err := gojwt.ParseWithClaims(jwtToken, &Claims{}, alg)
 	if err != nil {
-		return "", err
+		return Principal{}, err
 	} else if !token.Valid {
-		return "", err
+		return Principal{}, err
 	}
 
 	claims, ok := token.Claims.(*Claims)
 	if !ok {
-		return "", fmt.Errorf("unable to convert claims to standard claims")
+		return Principal{}, fmt.Errorf("unable to convert claims to standard claims")
 	}
 
-	return chronograf.Principal(claims.Subject), nil
+	return Principal{
+		Subject: claims.Subject,
+		Issuer:  claims.Issuer,
+	}, nil
 }
 
 // Token creates a signed JWT token from user that expires at Now + duration
-func (j *JWT) Token(ctx context.Context, user chronograf.Principal, duration time.Duration) (string, error) {
+func (j *JWT) Token(ctx context.Context, user Principal, duration time.Duration) (string, error) {
 	// Create a new token object, specifying signing method and the claims
 	// you would like it to contain.
 	now := j.Now().UTC()
 	claims := &Claims{
 		gojwt.StandardClaims{
-			Subject:   string(user),
+			Subject:   user.Subject,
+			Issuer:    user.Issuer,
 			ExpiresAt: now.Add(duration).Unix(),
 			IssuedAt:  now.Unix(),
 			NotBefore: now.Unix(),