+++ ui/src/shared/actions/auth.js
@@ -1,0 +1,8 @@
+export function receiveAuth(auth) {
+  return {
+    type: 'AUTH_RECEIVED',
+    payload: {
+      auth,
+    },
+  };
+}